Updated Grain to support dynamic revision numbers in the theme properties, like 0.6.+
Concrete values still must be specified for the major and minor numbers of the Grain version, declarations like 0+, 0.+ or 0.6+, won't accepted by Grain. This is done on purpose to ensure that a theme always uses compatible Grain version. Grain revisions on the other hand (like 0.6.1, 0.6.2, etc.) must be always interchangeable.
Signed-off-by: Dmitriy Pavlenko <dmitriy.pavlenko@sysgears.com>