Suppress broken pipe exception produced by Jetty on page load
When reloading the page several times very quickly the "Broken pipe" I/O exception is fired. It is normal for Jetty to throw such exception because if we do not throw then an application may waste resources generating and writing more content that cannot be sent. So it seems to be impossible to suppress one with custom Jetty configuration. Now we handle it properly in Grain code without writing noisy stack traces into log